requirements.txt
discord.py
python-dotenv
requests
flask
asyncio